refactor(cypress): move component utils + inspector into appBuilder; extract queryPanel
Continue the utils/ 3-folder taxonomy (appBuilder / marketplace / platform):
- appBuilder/components/ ← button, table, dropdown, datePicker (was datePickerWidget),
  listView (listviewWidget), multiSelect (multiselectWidget), modal, basicComponents
- appBuilder/inspector.js ← inspector barrel (internal import repointed to ./inspectorTree)
- appBuilder/querymanager/queryPanel.js ← NEW: extracted resizeQueryPanel out of
  marketplace/datasources/dataSource.js (it's a query-panel op, not a datasource op)
All import references updated (0 dangling; webpack-resolution smoke test green over
all 10 new paths). No behavioural changes.
