+++ cypress-tests/cypress/e2e/happyPath/appbuilder/commonTestcases/newSuits/table/tableCsa.cy.js
@@ -16,9 +16,9 @@ import {
   makeAllColumnsEditable,
   editTableCell,
   addFilter,
-} from "Support/utils/table";
+} from "Support/utils/appBuilder/components/table";
 import { openEditorSidebar, openAccordion } from "Support/utils/commonWidget";
-import { resizeQueryPanel } from "Support/utils/marketplace/datasources/dataSource";
+import { resizeQueryPanel } from "Support/utils/appBuilder/querymanager/queryPanel";
 
 // ---------------------------------------------------------------------------
 // Chunk 2 — Table component-specific actions (CSAs).
