+++ cypress-tests/cypress/e2e/happyPath/platform/commonTestcases/apps/inspectorOperations/globalsNodeHappyPath.cy.js
@@ -1,5 +1,5 @@
 import { fake } from "Fixtures/fake";
-import { navigateAndVerifyInspector } from "Support/utils/inspector";
+import { navigateAndVerifyInspector } from "Support/utils/appBuilder/inspector";
 
 describe("Globals - Inspector", () => {
   beforeEach(() => {
