+++ cypress-tests/cypress/support/utils/appBuilder/inspector.js
@@ -1,4 +1,4 @@
 // Barrel — re-exports everything that was previously defined here.
-// Function bodies have moved to ./appBuilder/inspectorTree.js.
+// Function bodies have moved to ./inspectorTree.js.
 // All 31 existing import sites continue to resolve without change.
-export * from "./appBuilder/inspectorTree";
+export * from "./inspectorTree";
